refactor(wire): split runtime observability providers
